backend/careminder/requirements.txt
django
djangorestframework-simplejwt
django-cors-headers
django-filter
djangorestframework
gunicorn
psycopg2-binary